Runway ends now have a hotspot which allows to select them as start position similar to parking
Corrected display for flight plan related tooltip messages.
Added hint for hidden objects in options dialog on page map display airport.
Changed fetch of cached parking, helipads and runways to avoid copying.
Hotspot for runway end right-click is a black and white circle.

+++ src/mappainter/mappaintermark.cpp
@@ -1110,9 +1110,9 @@ void MapPainterMark::paintCompassRose()
 
     QList<ageo::Pos> endpoints;
     QList<QPointF> endpointsScreen;
-    for(float angle = 0.f; angle < 360.f; angle += 5)
+    for(int angle = 0; angle < 360; angle += 5)
     {
-      endpoints.append(pos.endpoint(radiusMeter, magVar + angle));
+      endpoints.append(pos.endpoint(radiusMeter, magVar + static_cast<float>(angle)));
       endpointsScreen.append(wToSF(endpoints.constLast()));
     }
